Make user can work only with his own contacts

+++ src/services/contacts.js
@@ -2,17 +2,16 @@ import { ContactsCollection } from '../db/models/contacts.js';
 import { calculatePaginationData } from '../utils/calculatePaginationData.js';
 import { SORT_ORDER } from '../constants/index.js';
 
-export const getAllContacts = async ({
-  page,
-  perPage,
-  sortOrder = SORT_ORDER.ASC,
-  sortBy = '_id',
-  filter = {},
-}) => {
+export const getAllContacts = async (
+  { page, perPage, sortOrder = SORT_ORDER.ASC, sortBy = '_id', filter = {} },
+  userId,
+) => {
   const limit = perPage;
   const skip = (page - 1) * perPage;
 
-  const contactsQuery = ContactsCollection.find();
+  const contactsQuery = ContactsCollection.find()
+    .where('userId')
+    .equals(userId);
 
   if (filter.type) {
     contactsQuery.where('contactType').equals(filter.type);
@@ -22,9 +21,9 @@ export const getAllContacts = async ({
     contactsQuery.where('isFavourite').equals(filter.isFavourite);
   }
 
-  if (filter.userId) {
-    contactsQuery.where('userId').equals(filter.userId);
-  }
+  // if (filter.userId) {
+  //   contactsQuery.where('userId').equals(filter.userId);
+  // }
 
   const contactsCount = await ContactsCollection.find()
     .merge(contactsQuery)
@@ -44,8 +43,10 @@ export const getAllContacts = async ({
   };
 };
 
-export const getContactById = async (contactId) => {
-  const contacts = await ContactsCollection.findById(contactId);
+export const getContactById = async (contactId, userId) => {
+  const contacts = await ContactsCollection.findById(contactId)
+    .where('userId')
+    .equals(userId);
   return contacts;
 };
 
@@ -55,7 +56,12 @@ export const createContact = async (payload, userId) => {
   return contact;
 };
 
-export const updateContact = async (contactId, payload, options = {}) => {
+export const updateContact = async (
+  contactId,
+  payload,
+  userId,
+  options = {},
+) => {
   const rawResult = await ContactsCollection.findOneAndUpdate(
     { _id: contactId },
     payload,
@@ -64,7 +70,9 @@ export const updateContact = async (contactId, payload, options = {}) => {
       includeResultMetadata: true,
       ...options,
     },
-  );
+  )
+    .where('userId')
+    .equals(userId);
 
   if (!rawResult || !rawResult.value) return null;
 
@@ -74,10 +82,12 @@ export const updateContact = async (contactId, payload, options = {}) => {
   };
 };
 
-export const deleteContact = async (contactId) => {
+export const deleteContact = async (contactId, userId) => {
   const contact = await ContactsCollection.findOneAndDelete({
     _id: contactId,
-  });
+  })
+    .where('userId')
+    .equals(userId);
 
   return contact;
 };
